Add single-file checkpoint support and device selection
- Add support for loading .safetensors checkpoint files using StableDiffusionXLPipeline.from_single_file()
- Implement pipeline caching to avoid reloading when extracting multiple components
- Extract UNet, VAE, and text encoders from loaded pipeline
- Add --device CLI argument for explicit device selection (cuda, cpu, mps, etc.)
- Propagate device parameter through all pipeline loading calls to prevent incorrect CUDA autocast warnings
- Fix dtype mismatch by converting pixel_values to VAE's dtype before encoding
- Add test convenience script for local testing

All 38 tests passing.

+++ src/lora_trainer/train_loop.py
@@ -139,8 +139,8 @@ def train(
             if use_real_diffusion:
                 # Real diffusion training
                 with torch.no_grad():
-                    # Encode images to latents
-                    latents = vae.encode(pixel_values).latent_dist.sample()
+                    # Encode images to latents (convert to VAE's dtype)
+                    latents = vae.encode(pixel_values.to(vae.dtype)).latent_dist.sample()
                     latents = latents * vae.config.scaling_factor
 
                     # Encode prompts
